Implement all commands by running its corresponding task
